move dropdown to its own component; revert changes with Controller component
I was conceptualizing this all as a top-down approach when really it needs to be bottom up
