+++ src/DashboardUI/src/data-contracts/DropdownOption.ts
@@ -1,0 +1,4 @@
+export interface DropdownOption {
+  value: string;
+  label: string;
+}
